Digest認証とGoogleOAuth2.0を同じディレクトリに移動。
